Upgraded Jackson from 1.x to 2.5.2 and split out HystrixMetricsPoller unit test
