+++ hystrix-contrib/hystrix-metrics-event-stream/src/main/java/com/netflix/hystrix/contrib/metrics/eventstream/HystrixMetricsPoller.java
@@ -15,8 +15,6 @@
  */
 package com.netflix.hystrix.contrib.metrics.eventstream;
 
-import static org.junit.Assert.*;
-
 import java.io.IOException;
 import java.io.StringWriter;
 import java.util.concurrent.Executors;
@@ -26,19 +24,15 @@ import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
 
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonGenerator;
 import com.netflix.hystrix.HystrixCollapserKey;
 import com.netflix.hystrix.HystrixCollapserMetrics;
-import org.codehaus.jackson.JsonFactory;
-import org.codehaus.jackson.JsonGenerator;
-import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.netflix.hystrix.HystrixCircuitBreaker;
-import com.netflix.hystrix.HystrixCommand;
-import com.netflix.hystrix.HystrixCommandGroupKey;
 import com.netflix.hystrix.HystrixCommandKey;
 import com.netflix.hystrix.HystrixCommandMetrics;
 import com.netflix.hystrix.HystrixCommandMetrics.HealthCounts;
@@ -186,7 +180,7 @@ public class HystrixMetricsPoller {
             HystrixCircuitBreaker circuitBreaker = HystrixCircuitBreaker.Factory.getInstance(key);
 
             StringWriter jsonString = new StringWriter();
-            JsonGenerator json = jsonFactory.createJsonGenerator(jsonString);
+            JsonGenerator json = jsonFactory.createGenerator(jsonString);
 
             json.writeStartObject();
             json.writeStringField("type", "HystrixCommand");
@@ -394,76 +388,6 @@ public class HystrixMetricsPoller {
             Thread thread = defaultFactory.newThread(r);
             thread.setName(MetricsThreadName);
             return thread;
-        }
-    }
-
-    public static class UnitTest {
-
-        @Test
-        public void testStartStopStart() {
-            final AtomicInteger metricsCount = new AtomicInteger();
-
-            HystrixMetricsPoller poller = new HystrixMetricsPoller(new MetricsAsJsonPollerListener() {
-
-                @Override
-                public void handleJsonMetric(String json) {
-                    System.out.println("Received: " + json);
-                    metricsCount.incrementAndGet();
-                }
-            }, 100);
-            try {
-
-                HystrixCommand<Boolean> test = new HystrixCommand<Boolean>(HystrixCommandGroupKey.Factory.asKey("HystrixMetricsPollerTest")) {
-
-                    @Override
-                    protected Boolean run() {
-                        return true;
-                    }
-
-                };
-                test.execute();
-
-                poller.start();
-
-                try {
-                    Thread.sleep(500);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-
-                int v1 = metricsCount.get();
-
-                assertTrue(v1 > 0);
-
-                poller.pause();
-
-                try {
-                    Thread.sleep(500);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-
-                int v2 = metricsCount.get();
-
-                // they should be the same since we were paused
-                assertTrue(v2 == v1);
-
-                poller.start();
-
-                try {
-                    Thread.sleep(500);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-
-                int v3 = metricsCount.get();
-
-                // we should have more metrics again
-                assertTrue(v3 > v1);
-
-            } finally {
-                poller.shutdown();
-            }
         }
     }
 }
